+++ client/src/app/app-routing.module.ts
@@ -1,6 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { AdminAuthGuard, AuthenticatedGuard } from './core/guards/auth.guard';
+import {
+  AdminAuthGuard,
+  AuthenticatedGuard,
+  OwnerAuthGuard,
+} from './core/guards/auth.guard';
 import { HomeComponent } from './home.component';
 
 const routes: Routes = [
@@ -29,6 +33,14 @@ const routes: Routes = [
     loadChildren: () =>
       import('./tagging-interface/tagging-interface.module').then(
         (m) => m.TaggingInterfaceModule
+      ),
+  },
+  {
+    path: 'owner',
+    canActivate: [OwnerAuthGuard],
+    loadChildren: () =>
+      import('./owner-dashboard/owner-dashboard.module').then(
+        (m) => m.OwnerDashboardModule
       ),
   },
 ];