+++ client/src/app/core/guards/auth.guard.ts
@@ -50,3 +50,25 @@ export class AuthenticatedGuard implements CanActivate {
     return true;
   }
 }
+
+@Injectable()
+export class OwnerAuthGuard implements CanActivate {
+  constructor(private authService: AuthService, private router: Router) {}
+
+  /**
+   * Only allows the user if they are an admin.
+   */
+  canActivate(
+    _route: ActivatedRouteSnapshot,
+    _state: RouterStateSnapshot
+  ): boolean {
+    // If no user present, does not have access, redirect to login
+    if (!this.authService.isAuthenticated()) {
+      this.router.navigate(['/auth']);
+      return false;
+    }
+
+    // Check to make sure the user is an owner
+    return this.authService.user.roles.owner;
+  }
+}