+++ server/src/controllers/api/user.controller.spec.ts
@@ -3,6 +3,7 @@ import { UserService } from '../../services/user.service';
 import { User } from '../../schemas/user.schema';
 import { UserController } from './user.controller';
 import { RolesGuard } from '../../guards/role.guard';
+import { ConfigService } from '@nestjs/config';
 
 const testUser1: User = {
   _id: '1',
@@ -38,6 +39,10 @@ const userService = {
     return [testUser1, testUser2];
   },
 
+  async find(_id: string) {
+    return testUser1;
+  },
+
   addRole: jest.fn((_role: string, id: string) => {
     if (id == testUser1._id || id == testUser2._id) {
       return true;
@@ -62,6 +67,13 @@ const rolesGuard = {
   },
 };
 
+// Test config service
+const configService = {
+  get(_param: string) {
+    return 3;
+  },
+};
+
 describe('UserController', () => {
   // Controller being tested
   let userController: UserController;
@@ -73,6 +85,10 @@ describe('UserController', () => {
         {
           provide: UserService,
           useValue: userService,
+        },
+        {
+          provide: ConfigService,
+          useValue: configService,
         },
       ],
     })